changed to doubles

+++ libs/optimization_lib/include/objective_functions/AuxBendingNormal.h
@@ -22,7 +22,7 @@ private:
 	Eigen::Matrix< double, 6, 6> d2m_dNdN(int hi);
 	
 public:
-	float w1 = 1, w2 = 100, w3 = 100;
+	double w1 = 1, w2 = 100, w3 = 100;
 	OptimizationUtils::FunctionType functionType;
 	double planarParameter;
